ActiveTrain now handles new Intersection behaviour
Now the ActiveTrain does not try to change the active connection
of a track when it's route is changed.

This change was made to prevent other players being able to change
the route of another train in play, which is what would happen
previously if a route was changed.

Now a train will simply wait at the end of a track if it cannot
move to it's next intended track, and it will be up to the player
to change the track routing on the map to allow the train to move
again.

There is also a new boolean that dictates whether or not a route
can be changed to include a track that is an invalid connection.