Nudge track method now lives in mapCreator only
Having the nudge method in the actual model destroys a lot
of good assumptions you should have about the whereabouts
of a track, this can be very counterproductive when
manipulating intersections and discovering a track has
suddenly been displaced.

Nudging is redundant in the model in the sense that the code
knows that two tracks are no longer in the intersection, but
of course a user may assume two tracks are in the same intersection
just because they are touching.

Now nudging has the same effect to the map designer, helping
them to clarify where the orphan tracks are when breaking
an intersection, without all the mess in the rest of the code.

TL;DR: nudging lives on, doesn't mess with good code.